fix: move around cruise charts
